don't cast file length to int

+++ src/java/org/civilian/internal/asset/FileAsset.java
@@ -36,7 +36,7 @@ public class FileAsset extends AbstractAsset
 	public FileAsset(File file)
 	{
 		file_ = file;
-		setLength((int)file_.length());
+		setLength(file_.length());
 		setLastModified(file.lastModified());
 		
 		String name = file.getName();
